FAQ, Rule.. last edition

+++ src/main/java/com/fantasy/fantasyleague/FaqRule/Service/FAQServiceImpl.java
@@ -6,32 +6,45 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 // Needs refactoring, what if delete is not successful + how to determine it's not
 @Service
 public class FAQServiceImpl implements FAQService{
-    String deleteResponse = "Deleted Successfully";
-    String insertResponse = "Inserted Successfully";
-    String updateResponse = "Updated Successfully";
+    private String deleteResponse = "Deleted Successfully";
+    private String insertResponse = "Inserted Successfully";
+    private String updateResponse = "Updated Successfully";
+    private String deleteResponseF = "Unsuccessful Delete";
+    private String insertResponseF = "Unsuccessful Insert";
+    private String updateResponseF = "Unsuccessful Update";
 
     @Autowired
     FAQRepository faqRepository;
     @Override
     public String insertFAQ(FAQ faq) {
+        if(faqRepository.existsById(faq.getFaqID()))
+            return insertResponseF;
         faqRepository.save(faq);
         return insertResponse;
     }
 
     @Override
     public String updateFAQ(FAQ faq) {
-        faqRepository.deleteById(faq.getFaqID());
-        faqRepository.save(faq);
+        if(!faqRepository.existsById(faq.getFaqID()))
+            return updateResponseF;
+        FAQ faqToUpdate = faqRepository.getReferenceById(faq.getFaqID());
+        faqToUpdate.setAnswer(faq.getAnswer());
+        faqToUpdate.setDate(faq.getDate());
+        faqToUpdate.setQuestion(faq.getQuestion());
+        faqRepository.save(faqToUpdate);
         return updateResponse;
     }
 
     @Override
     public String deleteFAQ(FAQ faq) {
+        if(!faqRepository.existsById(faq.getFaqID()))
+            return deleteResponseF;
         faqRepository.deleteById(faq.getFaqID());
         return deleteResponse;
     }
@@ -40,4 +53,11 @@ public class FAQServiceImpl implements FAQService{
     public List<FAQ> getAllFAQ(){
         return faqRepository.findAll();
     }
+
+    @Override
+    public String deleteAllFAQ(){
+        faqRepository.deleteAll();
+        return deleteResponse;
+    }
+
 }
